Implement alternative variance definition
This consists in defining a different variance for each of the two equilibria of the model. Thus, it uses the previous calculation of the crossing times between equilibria so as to break the time series in pieces according to the equilibrium they correspond to.

This also adds support for plotting this alternative variance, including a numerical line for high temperature.